Simplify models
We don't need the Blob model, it was just overcomplicating things,
made things harder to understand, so the first model which is emitted
is YaraSearchResult, which simply contains the yara.Match with the
associated handler.

Instead of separate validate_chunk() and calculate_chunk() function,
we decided to keep only the calculate_chunk(), which is responsible
for validation too (if necessary), because there are formats, where
validation is not necessary at all, and there are formats, when
validation is not possible without calculating the chunk sizes.