Fix tar Chunk calculation
Use the tarfile in Python standard library instead of trying to implement
everything our own, because it can handle a lot more formats and can calculate
member sizes and positions correctly.
Instead of calculating the size, calculate the last member last byte's position
in the file with the help of tarfile.