Configure logger only when using the CLI
Unblob should work also as a library, we need to implement it in a way where
structlog can be configured by the caller side when used as a library.
This way the cli.py will not have a side effect and logging will be
configured only when calling the unblob command directly.