Implement CPIO format for all 4 versions
The CPIO format headers are very different for each version,
it's easier to treat them as separate formats and just
use common functions which are the same for all of them.

The formats:
  - ascii (magic: 07 07 01)
  - ascii with CRC (magic: 07 07 02)
  - old ASCII (magic: 07 07 07)
  - default / binary (magic: c7 c1)

The differences are:
  - Headers
  - Padding size (2 or 4 bytes)
  - calculating the filename size
  - calculating the file size

Handlers are implemented with subclasses, which makes sense in this case, only
the differences are implemented in the subclasses, everything else is common in
the base class.

Because of the structural typing, we can use classes instead of modules for Handlers.