Add RAR handler
It is using an external library to handle v4 and v5 format too.
It is using the unar external tool to extract found RAR files:
https://theunarchiver.com/command-line

The unar command will create empty files for password protected rar files and
empty directory without any files when it's encrypted. These are reflected in
the test cases.