Simplify header parsing and logging
Parsing a header is a very common operation, we can automate parsing and
logging with an extra optional attribute on the class.
Less boilerplate is necessary in subclasses when using this method.

The method is always loggging the parsed structure on DEBUG level, so we don't
need to think about an extra logger call.