Fix CPIO and TAR handlers and their respective test cases to support proper file padding.
CPIO: Add padding that could exists between the cpio trailer and the end-of-file (cpio aligns the file to 512 bytes).

TAR: the data in an archive is grouped into blocks, which are 512 bytes. Blocks are read and written in whole number multiples called records.
The number of blocks in a record (i.e., the size of a record in units of 512 bytes) is called the blocking factor.

We took the decision to pad tar chunks to tar's default blocking factor (20). Support for different blocking factor may be added in the future, if required.

Fix #58.