Modify file signature of LimitedStartReader
The aim here is to keep the possibility to parallelize handler
executions. For this, each hander needs to have the ability to be
exclusive with the given file, because they are doing file heavy
operations (seek, read) all over the place.
With this modification, each LimitedStartReader will read its own file
descriptor instead of using the a common one. For each handler, we are
already passing a different LimitedStartReader instance, which is good.