Skip unhandled Exceptions when calculating Chunks
There can be all kinds of Exceptions when trying to calculate chunks, sometimes
we can't do anything about them and might not even real errors.

One concrete example we run into is finding the magic value in a C header file
and the file length was smaller than the actual header size, so header parsing failed.

We still emit an error log and move on.