Fix UBIHandler
The original problem was that find_first changed the pointer,
and we were seeking by the wrong offset. If we seek to an absolute
position, it doesn't matter how find_first changed the pointer.

Test has been added for UBI.

Here is an article how to create UBI files
https://bootlin.com/blog/creating-flashing-ubi-ubifs-images/

The test UBI file is created with this config file:

    [rootfs_volume]
    mode=ubi
    image=apple1.txt
    vol_id=1
    vol_type=static
    vol_name=apple
    vol_alignment=1

    [rwdata_volume]
    mode=ubi
    image=cherry1.txt
    vol_id=2
    vol_type=dynamic
    vol_name=data
    vol_alignment=1
    vol_flags=autoresize

Then we ran this command:
$ ubinize -vv -o fruits.ubi -p 1024 -m 8 ubiconfig.cfg