Directly use shutil.COPY_BUFSIZE instead of a copy-paste value.
We should use these values directly from the standard library, instead
of having our copy of it, it is more "future proof" as well as can work
with different operation systems smoothly.

Its type is ignored by pyright, because COPY_BUFSIZE is not in the
shutil's __all__ directly - but it should not be a problem to use it
anyways, it is not considered private.

https://github.com/microsoft/pyright/blob/main/docs/typed-libraries.md#library-interface