Make unblob exit with appropirate exit code
Now there is a new EXIT_CODE variable to be able to set exit code from
anywhere in unblob, without immediately exiting.

These kind of global state managing resources should be located in the
new state.py. In the future, we want to add parallelism, and we need to
be aware for these, so the different processes / threads could share
these state resources.