Cache Git LFS directory
We used ~2Gb bandwidth of Git LFS in 1 day, because every GitHub action
is pulling the Git LFS objects unconditionally and with: lfs: true
doesn't cache pulled LFS objects by default.

By caching the .git/lfs directory it's possible to save that bandwith. The
invalidation logic is that if any of the Git LFS object hash changes or new
files added, it will pull it again.

The implementation is from: https://github.com/actions/checkout/issues/165

We only run these steps just before the test run, so if anything fails
before that, it should not pull anything.