Add support for LZO compression format.
The LZO file format holds a compressed size value in its header that we can use to
calculate the right end offset.

lzop utilities are a requirement to decompress carved out chunks, we added it to
Github Action tests.

To compress a file with lzop with highest level of compression (9):

```
lzop -9 -olorem.lzo lorem.txt
```

To compress a file with lzop with filter set (can be 1 to 16):

```
lzop --filter=10 -olorem.filter.lzo lorem.txt
```