Pass process_file the appropriate root directory
The "root" Path is used by process_files (by passing it to
make_extract_dir) to create the extraction directory for the
to-be-processed file.

Without this change, if the given "file" is in fact a directory, the
extraction path would contain the parent directory of the
to-be-extracted directory, which would not make sense.