Carve out unknown chunks to files + tests
Just put them in files the same way as we carve out ValidChunks,
the difference is the filename, it will be .unknown in this case,
because no Handler found it, and that they will not be extracted.

Only do any part of the extraction when there are some chunks.

Test file had to be added to the encrypted rar integration test,
as the __output__ will be different from now on.