Fix log formatting
Now there is a separate function for log formatting, which can be easily
unit tested.

Also, note for dealing with paths:
We can log paths either from the original files to be analyzed or from
the files created by us into the extraction directory. While it is a
really good thing that we log the later ones relative to the extraction
dir, we need to fully log the original files, as they can be anywhere.