Allow bigger cyclomatic complexity
We hit this multiple times already, not because we want to write terrible code,
but because it makes sense to have the main processing functions in one place
without too much indirection.
Even if these methods have higher complexity, they are throughly reviewed and
we think it worth the trade-off of having all logic in those bigger functions.

Also we have some bigger refactorings ahead of us, so these might be simpler.