Add support for LZIP compression format.
LZIP handler properly detects the LZIP magic and apply a naive approach derived
from the LZIP standard (https://datatracker.ietf.org/doc/html/draft-diaz-lzip)
to identify the end offset of identified chunk.

The idea is to read an 8 bytes uint64 every 2 bytes alignment (this is the alignment
of LZMA stream which is the central part of any LZIP compressed file) until we
read a value that corresponds to 'Member size'. We wrote the convert_int64 for this.

Member size is the 'total size of the member, including header and trailer.', so
when we read this value it _must_ correspond to our position in the file, minus
the start offset.

We use lziprecover as extraction binary because it's the only lzip extractor
that supports an explicit user controlled output file.