Handle EOFError in calculate_chunk
For ARC format, EOFError were triggered in cstruct when calling parse_header to
read past the end of the file, but other formats can have the same problem too,
so we handle the EOFError exception the same everywhere.  We still want to see
the error message during development (in verbose mode).