Refactor github workflows
Setup composite actions - These are a reusable workflow actions (like a function in a
programming language) which helps prepare the targeted environment.

This does every setup things before the actual workflows:
 - Install 3rd party applications with apt
 - Install pip + poetry
 - Install the virtual env with poetry
 - Git LFS pull - with cache

Later this can be extended for other setup activities, or optimize this
by doing caching for the whole virtualenv, etc.

Main workflow - This is the main workflow which runs for every push and
pull requests. It is doing the following jobs (in parallel):
 - check pre-commit
 - check pyright
 - run pytest