Add container public docker image workflow
We can build the docker image and push it to the GitHub's container
registry (ghcr.io) with the docker/build-push-action.

Since we are in heavy development, currently this worflow was set up to
run on main branch push events, and publish the container with the latest tag.

Before the actual push, we can build the container locally for testig.
It is using BuildKit, so multiple steps of the docker building are cached
automatically:
https://github.com/docker/build-push-action/blob/master/docs/advanced/test-before-push.md