Optimize file reading for lzma streams
Instead of reading the whole file into memory, we can feed it to
lzma.Decompressor chunk by chunk. It will set the .eof property when it reached
the end of the compressed stream. It will still read more, but those overflow
bytes will be kept in .unused_data property, so we can simply calculate where
it actually ended.
We make sure we only parsing lzma files, not xz with the lzma.FORMAT_ALONE
flag, because the standard library can handle both, but we have a separate
handler for xz.