Rust: adding Python extension module with packaging
As of now the extension packaging remains optional, unblob can be
installed without any Rust tooling (skipping the build of the
extension of course)

The extension will load into the `unblob._rust` namespace in
Python. Build uses the
[PyLimitedAPI](https://docs.python.org/3.8/c-api/stable.html)
functionality of Python, so that it doesn't need to get built for each
Python version. The generated binary will be forward compatible from
Python 3.8 onwards.

Packaging plugs into poetry's `build.py` feature. Unfortunately this
seams to be an undocumented extensibility point, but this is the only
way a native module can be built. It works by basically hooking into
a generated `setup.py` so that the `kwargs` sent to
`setuptools.setup()` can be overridden.