Fix logging when using as library
When unblob is used as a library, the structlog may already be
configured as the end-user likes. In this case:
    1) We don't need to configure it again with our CLI config.
    2) We still need to apply the "pretty_print_types" processor to the
processor chain, so our types will be printed nicely.
We add this to the begining of the processor chain, because
    - It is lightweight
    - At the end of the processor chain must be some render, which will
convert the whole event_dict to str, and usually before the renderer,
there are various amount of error / exception formatters.