Add support for Unix compress (LZW) compression format.
This module implements Unix compress'ed chunk identification.

We identify the end offset of any identified unix compress'ed chunk by performing
Lempel-Ziv-Welch decompression on a chunk starting from the identified start offset,
and ending at the end of the whole file being analyzed.

If we reach an invalid code or the stream ends in the middle of a code, we recursively
call the decompression, this time with a chunk ending at the failing position - 1.

Once the decompression procedure works without errors, that means we have a valid
chunk and can return its current end offset.

We use a small heuristic to return the right end offset. This heuristic tends
to work well when arbitrary data appended at the end of the stream is made of
random bytes (no repeating letters, no large set of ASCII letters).

It obviously can be wrong from time to time, leading to a compress'ed chunk
that we can decompress (obviously), but uncompressed data will contain garbage
bytes at the end.

Sadly, there is no way we can identify with 100% probability the end offset
of a compress'ed stream with byte precision if it is followed by other content.

The good news is that because of this behavior, it's highly unlikely we will
observe compress'ed chunks followed by other chunks in the wild. The only ones
I observed were followed by null bytes sentinels, which helps identifying the
exact end offset.