Add support for EXT filesystems.
EXT filesystem handlers support ext2, ext3, and ext4 with any (valid) block size. Test files have been generated to cover ext2, ext3, and ext4
with 1024, 2048, and 4096 block sizes.

In order to support ext4 we need a recent version of 7zip. We therefore in install the latest version of 7zip in Github actions.
We also had to fix calls to 7z on password protected archives since it changed in recent versions.