Remove `cache_logger_on_first_use` of structlog
This parameter means:
`wrap_logger` doesn't return an actual wrapped logger but a proxy that
assembles one when it's first used.  If this option is set to `True`,
this assembled logger is cached.

This caching mechanism's performance overhead is not that big
for us that we would need this optimization. However, it has its
disadvantages:
    1) Makes the testing of the CLI harder, as multiple usages of the CLI
would use the same logger object.
    2) Makes the logger object not pickleable, harder to use it in
multiprocessing.