tar: support smaller than RECORDSIZE archives without 2x zero blocks
There are tar files (such as what found in #118) which are not padded
to 20*BLOCKSIZE bytes and doesn't have 2*BLOCKSIZE ending
zeros.

The above cases are independent:

1. `tarfile.RECORDSIZE` is just tarfile.BLOCKSIZE times what GNU tar
calls as `--blocking-factor`. It's default value is 20 matching the
Python implementation, but it is not something we should expect to be
used and certainly. It can be smaller and even greater albeit it is
discouraged to increase it from the default.

2. According to `info tar` extractors shouldn't rely on the
existence of ending zero blocks:

> At the end of the archive file there are two 512-byte blocks filled
> with binary zeros as an end-of-file marker. A reasonable system
> should write such end-of-file marker at the end of an archive, but
> must not assume that such a block exists when reading an archive. In
> particular GNU 'tar' always issues a warning if it does not
> encounter it.

In reality GNU tar doesn't warn when a file without 2 zero blocks is
extracted.

Fixes: #118