Remove the recursion of `process_file` calls.
We agreed that we need to rework the recursion around the `process_file`
function, as this makes handling the unblob core codebase cumbersome to
deal with.

Now the `process_file` function calls are converted to `Task` objects,
which are put to a `Queue` and during unblobbing, we are getting
elements from the `Queue` and process them just like before.

Additionally, we need to think of multiprocessing. The only hard part is
to find the right way to recognize when there is nothing more to
process, and never will be - a right exit condition (because every file
processing can yield to multiple additional files to process). In this
matter, the multiprocessing.JoinableQueue can help us by manually
marking finished each Task in the Queue with the `task_done` function.
In this way the worker processes can use the blocking `queue.get`
indefinitely and the main process can check for the whole status by
`queue.join`.