Fix pyright typing issues with GzipReader
pyright has very specific rules how it detect private members, they are documented here:
https://github.com/microsoft/pyright/blob/main/docs/typed-libraries.md#library-interface

We don't want to copy-paste code from Python standard library to satisfy pyright, we
need to use the private methods, but there is no way to turn off these problems for
a block of code only in pyright, we can turn it off line-by-line or for a whole file,
so I isolated the questionable code to a separate module, so we can turn off the
issues for that file only.

This way it will easier to fix it in the future if the standard library breaks our code.