Improve LimitedFileReader to protect against offset size overflow.
File seeking APIs on Unix systems expects an offset of type off_t, which
is a signed 64 bits integer on modern systems. If we provide an offset
higher or lower than the upper and lower limits of that integer type,
Python will trigger a ValueError: "cannot fit 'int' into an offset-sized
integer".

As a defense against malformed chunks and mistakes in handler that may
have flown under the radar, we inserted an offset size check in
LimitedFileReader on calls to seek() and raised an InvalidInputFormat on
such cases.