Improve Chunk init error handling
Move the offset related validation inside the Chunk class, as a part of
the initialization. With this change, the caller does not need to do
these validations every time a Chunk is created.

Also modified the log level (error -> debug) for Chunk offset errors in
search_chunks_by_priority function. The reasoning behind that is we have
already created and handling the InvalidInputFormat error just for this
purpose: To skip an already calculated chunk of it has offset errors and
let the unblob proceed and find more appropriate Chunks to handle. This
is exactly the case here was well.