Add error reporting
During file processing, every encountered error will be reported
back to the main process through the `TaskResult` as some kind of
`Report`. This `Report` can be further extended with anything, however
now it only contains errors.

This removes the need to have a global state in the code, to exit code
of the CLI can be calculated by checking the reports in the end of
execution.