Do not attempt to extract encrypted file
The handlers should be able to detect whether a found ValidChunk is
password protected or not. If it is, the unblob should not even try to
extract it.

With this change, adding encryption detection to ZIP and RAR handlers.

Also removing the `extract_valid_chunk` function, because its new
functionality needs to be split:
    1) Carving out a valid chunk is moved to the new `carve_valid_chunk` function.
    2) Decision about further extracting the valid chunk is moved to the
Processor.