Fix result handling
For every process, we should create a `TaskResult` object in one place,
then mutate it (by adding reports or subsequent tasks) on-demand.

With the tasks, we were already doing this, however we need to do the
same with the reports to avoid loosing any report data for altering
between the "normal" processing and the exception processing.