Add own Pool for multiprocessing
We wanted to simplify how to multiprocessing of files are working for
unblob. The multiprocessing.Pool seemed like a great idea, however it
contains much (multithreading) overhead, which we just don't need.
We also wanted a way to be able to switch back to single processig
easily, because its advantages during development.

Based on these requirements, we have created our own MultiPool and
SinglePool classes, which are relying on a handler- and result callback
functions to operate.
Using Singlepool, the multiprocessing is not even envolved.
Esing MultiPool, we create underlying processes which are passing
around the tasks to be done and the results in multiprocessing queues,
using the definitive logic from multiprocessing.JoinableQueue.

The huge benefit from multiprocessing should be seen when the extracted
blob has many files, as multiple processes can work on different files
paralelly. This usually kicks in after the first few "rounds" of
extracting, as usually the first few blobs are big ones, exhausting its
handler process while the other processes are just waiting for new files
to appear for processing.