Improve ARC handler yara rule and processing.
Improve the ARC handler processing by returning early whenever we
encounter an invalid entry (name made of a null byte array). This tends
to happen a lot on our test corpus due to the loose yara rule.

Got rid of the unnecessary size, date, and time match in the ARC yara
rule given that we can't be 100% sure of the date/time format and that
the size allowed any four bytes value. Since then, we translated the
yara rule into the hexadecimal match rather than pure regex.