Move the "verbose" click.option to cli_options.py
Currently we have two CLI applications (unblob and unromfs extractor),
which are using the verbose option, both of them use it to configure the
logger with verbosity, so these CLI options must always look the same.
So instead of implementing it twice, just use them from a common
function.

Idea is based on this snippet:
https://github.com/pallets/click/issues/108#issuecomment-280489786