Fix overlapping valid chunks.
Due to the very nature of compression streams, we can't identify the
exact end offset of a stream. In some situations this can lead to a
valid compressed chunk overlapping the next chunk because the next chunk
starts with data that is valid to the decompressor (even though it will
be meaningless garbage when decompressed).

To overcome this situation, we introduce the 'fix_chunk_overlaps'
function that will browse through the list of valid chunks sorted by
increasing start_offset and adjust any overlapping chunk by setting its
end_offset to the next chunk start_offset value.