Fix JFFS2 EOFError handling.
When a JFFS2 filesystem is located at the very end of a file being
analyzed by unblob, the JFFS2 filesystem ended up not being properly
identified due to an uncaught EOFError exception being raised when
parsing a JFFS2 node header.

We fixed this by capturing this error early and breaking out of the loop
when that happens.

We also moved the node header validation to a dedicated function so that
cyclomatic complexity of calculate_chunk would stay low.