Improve LZH yara rule.
Instead of matching on the two first byte as "any byte", we use a list
of strings corresponding to every possible method id and use the 'any of
them' rule to match.

We set the YARA_MATCH_OFFSET to -2 given that method id sits 2 bytes
after the actual beginning of our LZH header.

This should be way more effective, especially since our method ids are
exactly four bytes long which corresponds to an atom lenght in yara.

We also offloaded the level_identifier validation to the yara rule and
made the python code simpler (no more constants, dropped a check).