pool: give worker processes a chance to exit gracefully
Instead of terminate, we place a sentinel object to the input queue so
that all processes have a chance to wake up and exit. In order to not
wait for arbitrary amount of time for the workers, we empty the input
queue from the main thread. In order to guarantee that this operation
is sound, we disallow placing subsequent tasks from other threads and
processes.

This change also means that the currently running task will always
finish.