Fixed squashfs padding (#237)
Squashfs files can be unpadded and 1k padded not just 4k padded.
There is no standard how squashfs should be padded, in the kernel it
depends on actually the kernel settings:
https://github.com/torvalds/linux/blob/5bfc75d92efd494db37f5c4c173d3639d4772966/fs/squashfs/Kconfig#L183
https://github.com/torvalds/linux/commit/7657cacf478940b995c2c73fdff981c13cc62c5ctorvalds/linux@7657cac
https://github.com/plougher/squashfs-tools/blob/23996602b0f7db407914aa9479fc41cb12b4c732/USAGE#L245
https://github.com/plougher/squashfs-tools/blob/62b7f5a411767409d2ec782a92ed632099b422d5/ACKNOWLEDGEMENTS#L101

Userspace tools usually use 4k padding as that is compatible with both
1k and 4k padding, but there could be 1k padded squashfs files out there:
https://dr-emann.github.io/squashfs/squashfs.html

For our end calculation using 4k could easily result in an overflow.
Unfortunately there is no easy way to detect this as the bytes_used is no
indication in this matter.

We are trying to use the longest padding as long as the padding consists
of zero bytes only. We can probably safely assume that no important data
starts with a 1k/4k zero bytes block.

Also fixed a problem when squashfs starts on a non-null offset.

Added a few unit test to cover chunk calculation and integration
test examples as well.