Fix ZIP handler end offset calculation.
This is a complete rewrite of the ZIP handler due to multiple issues
that have been identified during testing.

1. Python's built-in zipfile library is not reliable for our use.

This is due to the fact that zipfile seeks to the very end of the file
and then search backwards for the first occurence of central directory.

This is problematic because if we have a file fed to unblob that is made
of two concatenated ZIP files (with or without arbitrary content in
between), the end offset identified by zipfile would be the one of the
second zip chunk within the overall file, which is wrong in our case.

2. Simply finding the end of central directory marker is not enough

We previously searched for the EOCD marker and consider our first match
as the EOCD offset of the ZIP chunk we're currently investigating.

This is not sufficient because ZIP compression may not fully scramble
the data of the files it contains. This means that if you have a ZIP
within a ZIP, our first EOCD marker match will be the one of the ZIP
within the ZIP chunk we're currently trying to find the end offset of.

To overcome this, we iterate over all EOCD marker offsets within the
file and check for each of them whether or their offset corresponds to
the central directory offset + central directory size (both values are
read from the end of central directory entry). If offsets don't match,
that means our EOCD marker is not the right one.

This commit also get rid of the 'find_first' function as it turned into
dead code and vulture won't let me commit if I leave it there.