Fix LZMA handler yara rule and end offset computation.
We fixed the uncompressed size check in the LZMA yara rule as we were
checking the wrong side of a uint64 value using uint32s. This has now
been fixed and manually validated.

We fixed an issue in the LZMA handler that triggered the creation of
overlapping chunks (for example an LZMA stream followed by a squashfs
filesystem). Initially thought to be due to the inherent behavior of
decompression stream, we discovered that our initial reading loop was
reading one DEFAULT_BUFSIZE too much under specific circumstances.

We fixed the handlers in two ways, based on the uncompressed size field
value:

- if the uncompressed size is unknown, we wait for the 'eof' field of
our decompressor to become true, indicating that we've reached the
'end-of-stream marker'
- if the uncompressed size is known, we increment a counter and bail as
soon as we've uncompressed that amount of data

Integration tests for both situations have been added.