Add SIGINT and SIGTERM signal handling
The idea is to let us save the signal handlers and set up our own, then
if a signal is fired handle our things then call the original handlers.

As for the MultiPool:
    1. Not to propagate the signals to the worker processes to let the
main process handle the graceful closure.
    2. Make our implementation more robust by waiting for _SENTINEL not
just in the input queue, but in the output queue as well to wait for the
termination.
    3. At this point we can use multiprocessing.SimpleQueue instead of
the multiprocessing.Queue to leave out the threading shenanigans.