nix: dramatically reduce closure size of `unar`
`unar` depends on `gnustep-base` which pulls in a whopping 1.5GiB of
dependencies. There is an open issue about it in `nixpkgs` [1] which
haven't attracted much traction.

The workaround here is to override `gnustep-base` removing references
to the build dependencies from the output and while we are at it,
disabled some optional features not needed for `unar`.

`unar` is also overridden to reference our changed `gnustep-base`
derivation

---

[1]: https://github.com/NixOS/nixpkgs/issues/117208