Enhance squashfs extraction handling.
We've moved to our own version of squashfs-tools to be able to use the
'-no-exit-code' command line option. This option makes unsquashfs not
set the exit code to zero on non-fatal errors. non-fatal errors include
the inability to create special files such as block devices. With this
option, unblob will not report an error when such files cannot be
created due to missing super user privileges.

This is an opinionated change that follows the behavior of JFFS2
(jefferson) and RomFS (unromfs) extractors that simply do not attempt to
create special files if the process lacks super user privileges.

Dockerfile and github actions have been updated to include the
installation of our own fork of squashfs-tools.

An squashfs integration test files containing block device definition
has been added to check that unblob does not report an error.