Fix JFFS2 LZO integration tests.
With the addition of LZO compression support to Jefferson, we discovered
that our integration test cases for LZO were wrong.

What happened is that due to the lack of support for LZO, jefferson
simply copied the compressed content into the output file. We did not
catch that mistake during the early development of JFFS2 handlers.