Nix: force unblob to use an existing locale from the nix closure
Locale settings are environment dependent and Nix is happy to pick-up
global settings and work well as long as the glibc available in the
global environment uses compatible locale-archive format. However on
older systems it may not be the case and applications fall back to a
non-UTF-8 `C` locale, causing encoding issues in at least p7zip and
possible in other locations as well.

The `LOCALE_ARCHIVE_2_27` flag is specific [1] to the glibc used by
Nix, so if an external non-Nix based application is called, it will be
ignored and will fall back to the locale-archive available on the
system.

---

1: https://github.com/NixOS/nixpkgs/blob/d861c9c7e6051649a2a85f6666e1ce9e389302bb/pkgs/development/libraries/glibc/nix-locale-archive.patch