Fix XZ overall_size calculation.
The idea of the XZ handler is to identify the end of a chunk by finding
its footer with a string search. Then it validates that this chunk is
right by parsing the XZ records to calculate the overall_size value. If
both values correspond, we have a valid chunk.

There was an error in the overall_size calculation given that we
incremented by unpadded size of each record while we should increment by
the padded size of each record.

This commit fix this issue by incrementing by padded sizes.

An integration test file triggering the mis-calculation in the old code
has been created with the following command line:

xz -z --block-list=1021 -f /tmp/test.bin

This creates an XZ stream with small blocks of odd size, forcing the
creation of multiple padded blocks.

Fix #310