Fix gzip extraction failures due to null-bytes padding.
We initially carved out gzip chunks with their null byte padding, which
led to extraction error with 7z ("ERROR: There are some data after the
end of the payload data").

We no longer carve out the null byte padding, but let it be picked up as
unknown chunk and identify as padding later on.