cpio: convert YARA rule condition to handler checks and extend to all cpio formats
We optimized chunk matching to move some of the logic
into YARA rules and use smart conditions by checking
values relative to the specified magic using the @string
notation. However in YARA @string is actually a shortcut
for @string[1] which represents the address of the first
occurence of a given string. As in a YARA rule a given
string can match multiple times, but YARA evaluates all
string matches together there is no easy way in YARA to
return each string match AND check the conditions at the same
time. YARA rule either matches a file or not.

In our YARA rules using the @string match we always checked
the first occurence instead of checking it for each string,
which could either result in false-positive or false-negative
matches depending on the first string occurence.

In order to fix this inaccuracy we have to move the conditions
logic back into the handlers, which could potentially result in
a performance decrease.