Add false postivie checks to YAFFS handlers.
We discovered that YAFFS handlers (both yaffs and yaffs2) were
identifying false positive chunks in an httpd binary. This led to a
unknown chunk miscalculation, ultimately leading to a non-zero exit
code.

We fixed it by making strict header validation. We're checking that the
name is either full of null bytes or unicode decodable, and that both
checksum and sum_no_longer_used fields are set to 0xFFFF.