stop scanning file for chunk if a chunk matches till end of the file
As we are scanning the file in one pass due to having no priority level,
any further chunk will be inside of this last chunk. Chunks starting
before this chunk have already been matched. This way we can stop processing
even if the chunk is not covering the whole file, but matches till the end.