compress: handle decompression errors and too short chunks as error
We had a heuristic to recursively try to find the end of a valid compressed
chunk, however it yield to many false-positives. Unfortunately we have no
exact way to detect the end of the chunk as there is no end marker, or size
indication.

We just assume that the compression must be good and in most cases the
compressed content would be the whole file, so there is actually no need to
try to find the end recursively. We can handle extra 0byte padding after
the end, though the decompression would pick those up as well, which would
also result in some extra 0bytes in the decompressed output, but probably
this is the least bad in this scenario.

We also check if the chunk is at least 5 bytes long, as that is a sane minimum
for the header (3) + first data read (2).