Drop support for directory as input
The output directory structure is expected to reflect the input structure.

Having a single input file and a single extract directory is simple.

However supporting directories as input has problems:
subdirectories should be there in the output, but we also create
directories for extraction, thus it is very easy to craft an input that
has a conflicting output (usually an unblob output is one such input).