Simplify command output capture
A pair of "surrogateescape" decode and encode was used to get
from byte to bytes, with an internal utf-8 representation.
(read more about surrogateescape in https://peps.python.org/pep-0383/)
However the utf-8 representation was not used at all.

Although this simplification is not strictly necessary
(the results before and after must be the same) it took
some time reading it and understanding how the encoding works
during handling a problem that manifested in a failed command run
(https://github.com/IoT-Inspector/unblob/issues/356)