JSON output: handle bytes, and disregard JSON encoding failures
Quite unexpectedly, there is a kind of report, that has content in
bytes: ExtractCommandFailedReport, as the stdout/stderr streams are
in bytes.

It has made the metadata reporting fail in one case, so the custom
JSON encoding has been changed to make sensible output when running
into an unknown object, but never fail.