processing: introduce `ProcessResult`
This object holds an association of `Task`s and their respective
`TaskResult`s (and of course `Reports`). With this change we are able
to reconstruct what report is for what task, and also what subtasks
are coming from what tasks after the processing is finished.

This is a useful basis for metadata support, as any additional
information can later be added as reports.