Complete rewrite of XZ handler.
In order to emit a single ValidChunk for concatenated XZ compression
streams, we converted the handler to use Hyperscan like we did with
bzip2 handler (see c97007735f985ecf3fbd0563f1102485dc2a5846).

The idea is that for a given match, we use Hyperscan to look for end of
stream markers after our current start offset. If we have a match, we
read the whole chunk in memory and attempt to decompress it. If
decompression is successful, we adjust start and end offsets of our
context and instruct Hyperscan to continue looking for other end of
stream marker matches. Otherwise, we stop the search and return.